+++ Weather/Client/Services/Interfaces/IIconChooser.cs
@@ -5,6 +5,6 @@ namespace Weather.Client.Services
     public interface IIconChooser
     {
 
-        public string GetIcon(bool isDay, int cloudcover, float rain, float snow);
+        public string GetIcon(bool isDay, float cloudcover, float rain, float snow);
     }
 }
